+++ src/main/java/edu/ucsb/cs156/gauchoride/interceptors/RoleAdminDriverInterceptorConfiguration.java
@@ -1,0 +1,19 @@
+package edu.ucsb.cs156.gauchoride.interceptors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Component
+public class RoleAdminDriverInterceptorConfiguration implements WebMvcConfigurer {
+
+    @Autowired
+    private RoleAdminDriverInterceptor roleAdminDriverInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(roleAdminDriverInterceptor);
+    }
+
+}
